feat(player): add PlaybackRestored event
This commit introduces a new `MediaEvent.PlaybackRestored` event.

This event is triggered when playback is successfully restored in `MusicService`, either after queue restoration or when the player is ready after initialization.

The `AbsMainActivityFragment` now listens for `MediaEvent.PlaybackRestored`. If the `playOnStartupMode` preference is set to `WITH_EXPANDED_PLAYER`, the main activity's panel will be expanded when this event is received.

A new `proposeEvent` method has been added to `MediaEventBus` to allow for non-suspending event emission, which is used for the `PlaybackRestored` event.