fix queue position bug when swiping album cover to change songs
when " viewPager.addOnPageChangeListener(this)" is in onStart it can have an issue where the queue moves by multiple songs instead of a single song when swiping on the album cover in the player after opening app from background. this fixes it by moving it from onStart which gets called during the onResume lifecycle and puts it into the setupEventObserver which is only called during the onViewCreated method.